# Case1/FrontEnd/src/FrontEnd/Agents/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Cursusadministratie service","termsOfService":"None"},"basePath":"/","paths":{"/api/v1/Cursus":{"get":{"tags":["Cursus"],"operationId":"ApiV1CursusGet","consumes":[],"produces":["text/plain","application/json","text/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/CursusInstantie"}}}},"deprecated":false},"post":{"tags":["Cursus"],"operationId":"ApiV1CursusPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"cursus","in":"body","required":false,"schema":{"type":"array","items":{"$ref":"#/definitions/CursusInstantie"}}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/PostSuccess"}},"400":{"description":"Client Error","schema":{"$ref":"#/definitions/PostFailure"}}},"deprecated":false}},"/api/v1/Cursus/{id}":{"get":{"tags":["Cursus"],"operationId":"ApiV1CursusByIdGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/CursusInstantie"}}},"deprecated":false}}},"definitions":{"CursusInstantie":{"required":["cursus","startdatum"],"type":"object","properties":{"id":{"format":"int32","type":"integer"},"cursus":{"$ref":"#/definitions/Cursus"},"startdatum":{"pattern":"^[1-3]?[0-9]/1?[0-9]/\\d{4}$","type":"string"}}},"Cursus":{"required":["code","titel","duur"],"type":"object","properties":{"id":{"format":"int32","type":"integer"},"code":{"maxLength":10,"minLength":0,"type":"string"},"titel":{"maxLength":300,"minLength":0,"type":"string"},"duur":{"format":"int32","maximum":5,"minimum":1,"type":"integer"}}},"PostSuccess":{"type":"object","properties":{"inserted":{"format":"int32","type":"integer"},"total":{"format":"int32","type":"integer"}}},"PostFailure":{"type":"object","properties":{"errorCode":{"type":"string"},"errorMessage":{"type":"string"}}}},"securityDefinitions":{}}